+++ src/js/components/Peak/index.js
