+++ src/js/components/PeaksList/Peak.jsx
@@ -3,10 +3,8 @@ import { Link } from 'react-router-dom';
 import styles from './Peak.module.css';
 
 const Peak = ({ peak }) => {
-  // TODO change for proper address
-  const id = peak._links.self.href.split('http://localhost:8080/api/peaks/')[1];
   return (
-    <Link className={styles.card} to={'/peaks/'.concat(id)}>
+    <Link className={styles.card} to={`/peaks/${peak.id}`}>
       <img className={styles.photo} src={peak.photo} alt="mountain" />
       <div className={styles.peakName}>{peak.name}</div>
     </Link>
