+++ src/js/views/Peak/index.js
